refactor: create separate files for iTunes services
