+++ app/services/repoApi.js
@@ -1,0 +1,4 @@
+import { generateApiClient } from '@utils/apiUtils';
+const repoApi = generateApiClient('github');
+
+export const getRepos = (repoName) => repoApi.get(`/search/repositories?q=${repoName}`);
