+++ app/services/tests/repoApi.test.js
@@ -1,0 +1,19 @@
+import MockAdapter from 'axios-mock-adapter';
+import { getApiClient } from '@utils/apiUtils';
+import { getRepos } from '../repoApi';
+
+describe('repoApi tests', () => {
+  const repositoryName = 'mac';
+  it('should make the api call to "/search/repositories?q="', async () => {
+    const mock = new MockAdapter(getApiClient('github').axiosInstance);
+    const repoData = [
+      {
+        totalCount: 1,
+        items: [{ repositoryName }]
+      }
+    ];
+    mock.onGet(`/search/repositories?q=${repositoryName}`).reply(200, repoData);
+    const res = await getRepos(repositoryName);
+    expect(res.data).toEqual(repoData);
+  });
+});
